refactor(subscribe): move to src
