+++ packages/reactive/packages/decorator/subscribe/src/index.ts
@@ -1,0 +1,1 @@
+export { Subscribe } from './reactive-decorator.subscribe.func';
